@@ -58,7 +58,7 @@ build-backend = "poetry_dynamic_versioning.backend"                          # o
 [tool.poetry.dependencies] # must download to run
 # NOTE this setuptools dependency should be removed altogether once meson fortran builds are working
 setuptools = "<70"     # Required for any fortran builds until meson working
-python = ">=3.10.0"    # mandatory to declare the required python version
+python = ">=3.11.0"    # mandatory to declare the required python version
 matplotlib = ">=3.7.0" # Base requirement works, version specific to test outputs
 # Note netcdf <1.7.0 didn't seg fault if you did not properly close netcdf files.
 # >= 1.7.0 reasonably DOES seg fault without properly closing netcdf files.
