@@ -11,6 +11,7 @@ Other models defined here validate field types within child plugin models.
 import keyword
 import logging
 from typing import Union, Tuple
+import warnings
 
 # Third-Party Libraries
 from pydantic import (
@@ -37,8 +38,8 @@ ColorType = Union[ColorTuple, str]
 class PrettyBaseModel(BaseModel):
     """Make Pydantic models pretty-print by default.
 
-    This model overrides the default string representation of Pyantic models to generate
-    a user-friendly, JSON-formatted output with two-space indentation.
+    This model overrides the default string representation of Pydantic models to
+    generate a user-friendly, JSON-formatted output with two-space indentation.
     """
 
     def __str__(self) -> str:
@@ -146,7 +147,7 @@ def get_interfaces() -> set[str]:
     """Return a set of distinct interfaces.
 
     This function returns all available plugin interfaces. The results are cached for
-    runtime memory optimizaiton.
+    runtime memory optimization.
 
     Returns
     -------
@@ -223,7 +224,7 @@ class PluginModel(FrozenModel):
                 f"Must be one of {get_interfaces()}"
             )
         # the above exception handling would be further improved by checking the
-        # existence of plugin registry in the fuutre issue #906
+        # existence of plugin registry in the future issue #906
         if "package" not in metadata:
             err_msg = (
                 "Metadata for '%s' workflow plugin must contain 'package' key."
@@ -244,8 +245,6 @@ class PluginModel(FrozenModel):
 
         Parameters
         ----------
-        cls : Type
-            PluginModel class.
         value :
             Value of the 'interface' field to validate.
 
@@ -261,7 +260,7 @@ class PluginModel(FrozenModel):
         """
         valid_interfaces = get_interfaces()
         if value not in valid_interfaces:
-            err_msg = f"Incorrect interface:'{value}'.Must be one of {valid_interfaces}"
+            err_msg = f"Invalid interface:'{value}'. Must be one of {valid_interfaces}"
             LOG.critical(err_msg, exc_info=True)
             raise ValueError(err_msg)
         return value
@@ -271,7 +270,7 @@ class PluginModel(FrozenModel):
         cls: type["PluginModel"], values: dict[str, str | int | float | None]
     ):
         """
-        Set ``description`` to first line of ``dosctring`` field if not provided.
+        Set ``description`` to first line of ``docstring`` field if not provided.
 
         Parameters
         ----------
@@ -312,10 +311,12 @@ class PluginModel(FrozenModel):
         str
             Validated ``description`` string.
 
-        Raises
-        ------
-        PydanticCustomError
+        Warns
+        -----
+        FutureWarning
             If the ``description`` field violates any of the validation rules.
+            This will raise a ValidationError in a future release.
+
         """
         error_messages = {
             "single_line": "Description must be a single line.",
@@ -324,27 +325,37 @@ class PluginModel(FrozenModel):
             ),
             "length_error": "Description cannot be more than 72 characters, reduce by:",
         }
-        if "\n" in value:
-            LOG.critical(
-                "'error': %s 'input_provided': %r",
-                error_messages["single_line"],
-                value,
-                exc_info=True,
+        try:
+
+            def _log_and_raise_custom_pydantic_errors(
+                error_type: str, error_message: str, value: str
+            ):
+                warnings.warn(
+                    f"'error': {error_message} 'input_provided': {value}",
+                    FutureWarning,
+                    stacklevel=3,
+                )
+                raise PydanticCustomError(error_type, error_message)
+
+            if "\n" in value:
+                _log_and_raise_custom_pydantic_errors(
+                    "single_line", error_messages["single_line"], value
+                )
+            elif not value or not value[0].isalnum() or not value.endswith("."):
+                _log_and_raise_custom_pydantic_errors(
+                    "format_error", error_messages["format_error"], value
+                )
+            elif len(value) > 72:
+                excess_length = len(value) - 72
+                err_msg = f"{error_messages['length_error']} {excess_length} characters"
+                _log_and_raise_custom_pydantic_errors("length_error", err_msg, value)
+
+        except PydanticCustomError as e:
+            warnings.warn(
+                f"Future ValidationError encountered. This current warning will become "
+                f"an error in a future release. {e}",
+                FutureWarning,
+                stacklevel=2,
             )
-            raise PydanticCustomError("single_line", error_messages["single_line"])
-        if not (value[0].isalnum() and value.endswith(".")):
-            LOG.critical(
-                "'error': %s 'input_provided': %r",
-                error_messages["format_error"],
-                value,
-                exc_info=True,
-            )
-            raise PydanticCustomError("format_error", error_messages["format_error"])
-        if len(value) > 72:
-            excess_length = len(value) - 72
-            err_msg = f"{error_messages['length_error']} {excess_length} characters"
-            LOG.critical(
-                "'error': %s 'input_provided': %r", err_msg, value, exc_info=True
-            )
-            raise PydanticCustomError("length_error", err_msg)
+
         return value
